Factorize the `disp_vert` functions into <CGAL/Mesh_3/utilities.h>
