+++ Mesh_3/include/CGAL/Mesh_3/utilities.h
@@ -27,12 +27,41 @@
 
 #include <CGAL/license/Mesh_3.h>
 
+#include <CGAL/Has_timestamp.h>
+#include <iterator>
+#include <string>
+#include <sstream>
 
 namespace CGAL {
 
 namespace Mesh_3 {
 namespace internal {
   
+struct Debug_messages_tools {
+  template <typename Vertex_handle>
+  static std::string disp_vert(Vertex_handle v, Tag_true) {
+    std::stringstream ss;
+    ss.precision(17);
+    ss << (void*)(&*v) << "[ts=" << v->time_stamp() << "]"
+       << "(" << v->point() <<")";
+    return ss.str();
+  }
+
+  template <typename Vertex_handle>
+  static std::string disp_vert(Vertex_handle v, Tag_false) {
+    std::stringstream ss;
+    ss.precision(17);
+    ss << (void*)(&*v) << "(" << v->point() <<")";
+    return ss.str();
+  }
+
+  template <typename Vertex_handle>
+  static std::string disp_vert(Vertex_handle v)
+  {
+    typedef typename std::iterator_traits<Vertex_handle>::value_type Vertex;
+    return disp_vert(v, CGAL::internal::Has_timestamp<Vertex>());
+  }
+};
 
 /**
  * @class First_of
